load_db/src/app/requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ipykernel==5.4.3
ipython-genutils==0.2.0
ipython==7.20.0
jedi==0.18.0
jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
markupsafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0; os_name != 'nt'
pycparser==2.20
pygments==2.7.4
pymysql==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyzmq==22.0.2
send2trash==1.5.0
six==1.15.0
sqlalchemy==1.3.23
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
